Fixed visibility of incomplete bugs, ie: bugs with no associated user
Since we are storing the bug of anon user and then we assign the user to the bug when he logs in. We might have some cases when user do not complete the loggin process, that generates bugs with no users, we can call them "incomplete" bugs.
I added two methods to the manager of the Issue Class: incomplete and complete to filter issues by state.
These two methos now are used on the views to filter the bugs and only show the complete ones.