feat: Add functionality to add domain to company in domain.html
The code changes in this commit add functionality to the domain.html template to allow users to add a domain to a company. If the domain does not already belong to a company, a form is displayed where the user can enter the company name and submit it to add the domain to the company. This feature enhances the user experience and improves the organization of domains within companies.